==> notebooks/2. Scrapping Worten.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import lxml.html as lh

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import re
import time
from datetime import datetime
import matplotlib.dates as mdates
import matplotlib.ticker as ticker
from urllib.request import urlopen

#SS
#url = 'https://www.worten.es/productos/tv-video-y-sonido/televisores/de-samsung?diagonal-de-la-pantalla=65_27__27_%20(165%20cm)&diagonal-de-la-pantalla=60_27__27_%20(152%20cm)&resolucion=4K%20Ultra%20HD&tecnologia-del-fabricante=LED'
#LG
url = 'https://www.worten.es/productos/tv-video-y-sonido/televisores/de-lg?diagonal-de-la-pantalla=65_27__27_%20(165%20cm)&diagonal-de-la-pantalla=60_27__27_%20(152%20cm)&resolucion=4K%20Ultra%20HD&tecnologia-del-fabricante=LED'
html = requests.get(url).content
soup = BeautifulSoup(html, 'lxml')
table = soup.find_all('div', {'class': ['w-product-list__row']})[0]
table

rows = table.find_all(['div','h3','span'], {'class' : ['w-product__title','w-currentPrice iss-current-price']})
rows_parsed = [row.text for row in rows]
rows_parsed

row_split = 2
rows_refactored = [rows_parsed[x:x+row_split] for x in range(0, len(rows_parsed), row_split)]
rows_refactored

df = pd.DataFrame(rows_refactored, columns=['Model', 'Price'])
df

#df.to_csv('../data/worten_scrap/LG/worten_0511_LG.csv')

df.to_csv('../data/worten_scrap/SS/worten_0511_SS.csv')
