==> notebooks/4. WORTEN.ipynb <==
## LG SCRPAING DATASET

import numpy as np
import pandas as pd

df_0310 = pd.read_csv('../data/worten_scrap/LG/worten_0310_LG_OK.csv')
df_0510 = pd.read_csv('../data/worten_scrap/LG/worten_0510_LG_OK.csv')
df_0710 = pd.read_csv('../data/worten_scrap/LG/worten_0710_LG.csv')
df_0810 = pd.read_csv('../data/worten_scrap/LG/worten_0810_LG.csv')
df_1010 = pd.read_csv('../data/worten_scrap/LG/worten_1010_LG.csv')
df_1110 = pd.read_csv('../data/worten_scrap/LG/worten_1110_LG.csv')
df_1310 = pd.read_csv('../data/worten_scrap/LG/worten_1310_LG.csv')
df_1410 = pd.read_csv('../data/worten_scrap/LG/worten_1410_LG.csv')
df_1510 = pd.read_csv('../data/worten_scrap/LG/worten_1510_LG.csv')
df_1610 = pd.read_csv('../data/worten_scrap/LG/worten_1610_LG.csv')
df_1710 = pd.read_csv('../data/worten_scrap/LG/worten_1710_LG.csv')
df_1910 = pd.read_csv('../data/worten_scrap/LG/worten_1910_LG.csv')
df_2010 = pd.read_csv('../data/worten_scrap/LG/worten_2010_LG.csv')
df_2110 = pd.read_csv('../data/worten_scrap/LG/worten_2110_LG.csv')
df_2210 = pd.read_csv('../data/worten_scrap/LG/worten_2210_LG.csv')
df_2310 = pd.read_csv('../data/worten_scrap/LG/worten_2310_LG.csv')
df_2410 = pd.read_csv('../data/worten_scrap/LG/worten_2410_LG.csv')
df_2510 = pd.read_csv('../data/worten_scrap/LG/worten_2510_LG.csv')
df_2610 = pd.read_csv('../data/worten_scrap/LG/worten_2610_LG.csv')
df_2710 = pd.read_csv('../data/worten_scrap/LG/worten_2710_LG.csv')
df_2810 = pd.read_csv('../data/worten_scrap/LG/worten_2810_LG.csv')
df_0111 = pd.read_csv('../data/worten_scrap/LG/worten_0111_LG.csv')
df_0211 = pd.read_csv('../data/worten_scrap/LG/worten_0211_LG.csv')
df_0311 = pd.read_csv('../data/worten_scrap/LG/worten_0311_LG.csv')
df_0411 = pd.read_csv('../data/worten_scrap/LG/worten_0411_LG.csv')
#df_0511 = pd.read_csv('../data/worten_scrap/LG/worten_0511_LG.csv')
#df_0611 = pd.read_csv('../data/worten_scrap/LG/worten_0611_LG.csv')
#df_0711 = pd.read_csv('../data/worten_scrap/LG/worten_0711_LG.csv')
#df_0811 = pd.read_csv('../data/worten_scrap/LG/worten_0811_LG.csv')

df_0310['day']=1
df_0510['day']=2
df_0710['day']=3
df_0810['day']=4
df_1010['day']=5
df_1110['day']=6
df_1310['day']=7
df_1410['day']=8
df_1510['day']=9
df_1610['day']=10
df_1710['day']=11
df_1910['day']=12
df_2010['day']=13
df_2110['day']=14
df_2210['day']=15
df_2310['day']=16
df_2410['day']=17
df_2510['day']=18
df_2610['day']=19
df_2710['day']=20
df_2810['day']=21
df_0111['day']=22
df_0211['day']=23
df_0311['day']=24
df_0411['day']=25
#df_0511['day']=26
#df_0611['day']=27
#df_0711['day']=28
#df_0811['day']=29

df = pd.concat([df_0310,df_0510,df_0710,df_0810,df_1010,df_1110,
                              df_1310,df_1410,df_1510,df_1610,df_1710, df_1910, df_2010, df_2110, df_2210,
               df_2310, df_2410, df_2510, df_2610, df_2710, df_2810, df_0111, df_0211, df_0311, df_0411])

df['Price'] = df['Price'].str.replace("€", "-")
df['Price'] = df['Price'].str.replace(",", ".")
df['Price'] = df['Price'].map(lambda x: str(x)[:-1])
df['TVs'] = df['Price'].str.contains('-')
df.drop(df[df['TVs'] == True].index, inplace=True)
df = df.rename(columns={'Price':'Price €'})

WORTEN_LG = pd.DataFrame(df)
WORTEN_LG = WORTEN_LG.drop(columns=['Unnamed: 0','TVs'])
pd.to_numeric(WORTEN_LG['Price €'])
WORTEN_LG.to_csv('../data/worten_scrap/LG/WORTEN_LG.csv', index=False)

WORTEN_LG = pd.read_csv('../data/worten_scrap/LG/WORTEN_LG.csv')

WORTEN_LG = pd.pivot_table(WORTEN_LG, values='Price €', index=['Model'],
                      columns=['day'])

WORTEN_LG.to_csv('../data/worten_scrap/LG/WORTEN_LG_SCRAP.csv')

WORTEN_LG_SCRAP = pd.read_csv('../data/worten_scrap/LG/WORTEN_LG_SCRAP.csv')

# SAMSUNG SCRAPING DATASET

SS_0310 = pd.read_csv('../data/worten_scrap/SS/worten_0310_SS_OK.csv')
SS_0510 = pd.read_csv('../data/worten_scrap/SS/worten_0510_SS_OK.csv')
SS_0710 = pd.read_csv('../data/worten_scrap/SS/worten_0710_SS.csv')
SS_0810 = pd.read_csv('../data/worten_scrap/SS/worten_0810_SS.csv')
SS_1010 = pd.read_csv('../data/worten_scrap/SS/worten_1010_SS.csv')
SS_1110 = pd.read_csv('../data/worten_scrap/SS/worten_1110_SS.csv')
SS_1310 = pd.read_csv('../data/worten_scrap/SS/worten_1310_SS.csv')
SS_1410 = pd.read_csv('../data/worten_scrap/SS/worten_1410_SS.csv')
SS_1510 = pd.read_csv('../data/worten_scrap/SS/worten_1510_SS.csv')
SS_1610 = pd.read_csv('../data/worten_scrap/SS/worten_1610_SS.csv')
SS_1710 = pd.read_csv('../data/worten_scrap/SS/worten_1710_SS.csv')
SS_1910 = pd.read_csv('../data/worten_scrap/SS/worten_1910_SS.csv')
SS_2010 = pd.read_csv('../data/worten_scrap/SS/worten_2010_SS.csv')
SS_2110 = pd.read_csv('../data/worten_scrap/SS/worten_2110_SS.csv')
SS_2210 = pd.read_csv('../data/worten_scrap/SS/worten_2210_SS.csv')
SS_2310 = pd.read_csv('../data/worten_scrap/SS/worten_2310_SS.csv')
SS_2410 = pd.read_csv('../data/worten_scrap/SS/worten_2410_SS.csv')
SS_2510 = pd.read_csv('../data/worten_scrap/SS/worten_2510_SS.csv')
SS_2610 = pd.read_csv('../data/worten_scrap/SS/worten_2610_SS.csv')
SS_2710 = pd.read_csv('../data/worten_scrap/SS/worten_2710_SS.csv')
SS_2810 = pd.read_csv('../data/worten_scrap/SS/worten_2810_SS.csv')
SS_0111 = pd.read_csv('../data/worten_scrap/SS/worten_0111_SS.csv')
SS_0211 = pd.read_csv('../data/worten_scrap/SS/worten_0211_SS.csv')
SS_0311 = pd.read_csv('../data/worten_scrap/SS/worten_0311_SS.csv')
SS_0411 = pd.read_csv('../data/worten_scrap/SS/worten_0411_SS.csv')
#SS_0511 = pd.read_csv('../data/worten_scrap/SS/worten_0511_SS.csv')
#SS_0611 = pd.read_csv('../data/worten_scrap/SS/worten_0611_SS.csv')
#SS_0711 = pd.read_csv('../data/worten_scrap/SS/worten_0711_SS.csv')
#SS_0811 = pd.read_csv('../data/worten_scrap/SS/worten_0811_SS.csv')

SS_0310['day']=1
SS_0510['day']=2
SS_0710['day']=3
SS_0810['day']=4
SS_1010['day']=5
SS_1110['day']=6
SS_1310['day']=7
SS_1410['day']=8
SS_1510['day']=9
SS_1610['day']=10
SS_1710['day']=11
SS_1910['day']=12
SS_2010['day']=13
SS_2110['day']=14
SS_2210['day']=15
SS_2310['day']=16
SS_2410['day']=17
SS_2510['day']=18
SS_2610['day']=19
SS_2710['day']=20
SS_2810['day']=21
SS_0111['day']=22
SS_0211['day']=23
SS_0311['day']=24
SS_0411['day']=25
#SS_0511['day']=26
#SS_0611['day']=27
#SS_0711['day']=28
#SS_0811['day']=29

df_SS = pd.concat([SS_0310,SS_0510,SS_0710,SS_0810,SS_1010,SS_1110,
                              SS_1310,SS_1410,SS_1510,SS_1610,SS_1710, SS_1910, SS_2010, SS_2110,SS_2210,
                  SS_2310, SS_2410,SS_2510, SS_2610, SS_2710, SS_2810, SS_0111, SS_0211, SS_0311, SS_0411])

df_SS['Price'] = df_SS['Price'].str.replace("€", "-")
df_SS['Price'] = df_SS['Price'].str.replace(",", ".")
df_SS['Price'] = df_SS['Price'].map(lambda x: str(x)[:-1])
df_SS['TVs'] = df_SS['Price'].str.contains('-')
df_SS.drop(df_SS[df_SS['TVs'] == True].index, inplace=True)
df_SS = df_SS.rename(columns={'Price':'Price €'})

WORTEN_SS = pd.DataFrame(df_SS)
WORTEN_SS = WORTEN_SS.drop(columns=['Unnamed: 0','TVs'])
pd.to_numeric(WORTEN_SS['Price €'])
WORTEN_SS.to_csv('../data/worten_scrap/SS/WORTEN_SS.csv', index=False)

WORTEN_SS = pd.read_csv('../data/worten_scrap/SS/WORTEN_SS.csv')

WORTEN_SS = pd.pivot_table(WORTEN_SS, values='Price €', index=['Model'],
                      columns=['day'])

WORTEN_SS.to_csv('../data/worten_scrap/SS/WORTEN_SS_SCRAP.csv')

WORTEN_SS_SCRAP = pd.read_csv('../data/worten_scrap/SS/WORTEN_SS_SCRAP.csv')

worten_concat = pd.concat([WORTEN_LG_SCRAP, WORTEN_SS_SCRAP])

worten_concat.fillna(worten_concat.median(axis=0), inplace=True)

worten_concat.to_csv('../data/worten_scrap/WORTEN_CONCAT.csv')
